Add initial controlplane import contracts
