+++ tools/controlplane/src/controlplane_tool/cli/commands.py
@@ -1,94 +1,17 @@
 from __future__ import annotations
 
-from pathlib import Path
-
 import typer
 from pydantic import ValidationError
 
-from controlplane_tool.building.requests import BuildRequest
+from controlplane_tool.building.gradle_executor import GradleCommandExecutor
 from controlplane_tool.building.tasks import CommandExecutionResult
 from controlplane_tool.orchestation.flow_catalog import resolve_flow_definition
-from controlplane_tool.building.gradle_planner import (
-    build_gradle_command,
-    plan_module_matrix_commands,
-)
-from controlplane_tool.app.paths import default_tool_paths
 from controlplane_tool.orchestation.prefect_runtime import run_local_flow
-from controlplane_tool.core.shell_backend import SubprocessShell
 
 CLI_CONTEXT_SETTINGS = {
     "allow_extra_args": True,
     "ignore_unknown_options": True,
 }
-
-class GradleCommandExecutor:
-    def __init__(self, repo_root: Path | None = None) -> None:
-        self.repo_root = repo_root or default_tool_paths().workspace_root
-        self._shell = SubprocessShell()
-
-    def _build_command(
-        self,
-        action: str,
-        profile: str,
-        modules: str | None,
-        extra_gradle_args: list[str],
-    ) -> list[str]:
-        request = BuildRequest(
-            action=action,
-            profile=profile,
-            modules=modules,
-            extra_gradle_args=extra_gradle_args,
-        )
-        return build_gradle_command(
-            repo_root=self.repo_root,
-            request=request,
-            extra_gradle_args=extra_gradle_args,
-        )
-
-    def execute(
-        self,
-        action: str,
-        profile: str,
-        modules: str | None,
-        extra_gradle_args: list[str],
-        dry_run: bool,
-    ) -> CommandExecutionResult:
-        command = self._build_command(action, profile, modules, extra_gradle_args)
-        if dry_run:
-            return CommandExecutionResult(command=command, return_code=0, dry_run=True)
-
-        completed = self._shell.run(command, cwd=self.repo_root, dry_run=False)
-        return CommandExecutionResult(
-            command=command,
-            return_code=completed.return_code,
-            dry_run=False,
-            stdout=completed.stdout,
-            stderr=completed.stderr,
-        )
-
-    def execute_matrix(
-        self,
-        task: str,
-        modules: str | None,
-        max_combinations: int,
-        extra_gradle_args: list[str],
-        dry_run: bool,
-    ) -> tuple[list[list[str]], int]:
-        commands = plan_module_matrix_commands(
-            repo_root=self.repo_root,
-            task=task,
-            max_combinations=max_combinations,
-            modules_csv=modules,
-            extra_gradle_args=extra_gradle_args,
-        )
-        if dry_run:
-            return commands, 0
-
-        for command in commands:
-            completed = self._shell.run(command, cwd=self.repo_root, dry_run=False)
-            if completed.return_code != 0:
-                return commands, completed.return_code
-        return commands, 0
 
 
 def _combined_extra_gradle_args(
