+++ tools/controlplane/tests/test_command_resolver.py
@@ -1,4 +1,4 @@
-from controlplane_tool.core.command_resolver import CommandResolver
+from controlplane_tool.scenario.command_resolver import CommandResolver
 
 
 def test_replace_substitutes_single_placeholder() -> None:
